backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
google-cloud-firestore==2.11.1
python-dotenv==1.0.0
Pillow==10.0.1
pytesseract==0.3.10
numpy==1.24.3
opencv-python-headless==4.8.0.76
